requirements.txt
openai
streamlit
fastapi
uvicorn
python-dotenv
reportlab